Misc (re #1068): fixed various compilation warnings with Visual Studio 6 and 2005
git-svn-id: https://svn.pjsip.org/repos/pjproject/trunk@3262 74dad513-b988-da41-8d7b-12977e46ad98

+++ pjnath/src/pjnath/ice_strans.c
@@ -268,7 +268,7 @@ static pj_status_t add_update_turn(pj_ice_strans *ice_st,
 	/* Also if this component's default candidate is set to relay,
 	 * move it temporarily to something else.
 	 */
-	if (comp->default_cand == cand - comp->cand_list) {
+	if ((int)comp->default_cand == cand - comp->cand_list) {
 	    /* Init to something */
 	    comp->default_cand = 0;
 	    /* Use srflx candidate as the default, if any */
